+++ ssio-core/src/jvmAndAndroidMain/kotlin/dev/karmakrafts/ssio/nio/NioFileSink.kt
@@ -36,7 +36,7 @@ internal class NioFileSink(
     path: NioPath
 ) : AsyncRawSink {
     companion object {
-        private const val CHUNK_SIZE: Int = 8192
+        private const val CHUNK_SIZE: Int = 4096
     }
 
     init {